requirements.txt
fastapi==0.66.0
uvicorn[standard]==0.14.0
PyYAML==5.4.1
SQLAlchemy==1.4.20
psycopg2==2.9.1
alembic==1.6.5
Jinja2==3.0.1
pydantic==1.8.2
passlib[bcrypt]==1.7.4
email-validator==1.1.3
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
pytest==6.2.4
requests==2.26.0
user-agents==2.2.0